feat: remove old plan a lesson page and promote new one
