+++ src/__tests__/pages/lesson-planning.test.tsx
@@ -1,115 +1,153 @@
 import { screen } from "@testing-library/react";
 
-import PlanALesson from "../../pages/lesson-planning";
-import { CTA, PlanningPage } from "../../common-lib/cms-types";
+import PlanALesson, { getStaticProps } from "../../pages/lesson-planning";
 import renderWithProviders from "../__helpers__/renderWithProviders";
-import {
-  mockImageAsset,
-  mockSeo,
-  mockVideoAsset,
-  portableTextFromString,
-} from "../__helpers__/cms";
 import renderWithSeo from "../__helpers__/renderWithSeo";
 
-export const testPlanningPageData: PlanningPage = {
-  id: "01",
-  title: "Planning title",
-  heading: "Planning heading",
-  summaryPortableText: portableTextFromString("Planning summary"),
-  lessonElements: {
-    introQuiz: {
-      title: "Starter quiz title",
-      bodyPortableText: portableTextFromString("Intro quiz body"),
-    },
-    video: {
-      title: "Video title",
-      bodyPortableText: portableTextFromString("Video body"),
-    },
-    slides: {
-      title: "Slides title",
-      bodyPortableText: portableTextFromString("Slides body"),
-    },
-    worksheet: {
-      title: "Worksheet title",
-      bodyPortableText: portableTextFromString("Worksheet body"),
-    },
-    exitQuiz: {
-      title: "Exit quiz title",
-      bodyPortableText: portableTextFromString("Exit quiz body"),
-    },
-  },
-  lessonElementsCTA: {
-    label: "elements label",
-    linkType: "external",
-    external: "https://example.com",
-  },
-  stepsHeading: "steps",
-  steps: [
-    {
-      title: "step one",
-      bodyPortableText: portableTextFromString("step one body"),
-    },
-  ],
-  stepsCTA: {
-    label: "Steps CTA",
-    linkType: "internal",
-    internal: {
-      id: "homepage",
-      contentType: "homepage",
-    },
-  },
-  learnMoreHeading: "learn more heading",
-  learnMoreBlock1: {
-    title: "learn more block 1",
-    bodyPortableText: portableTextFromString("block 1 text"),
-    alignMedia: "left",
-    mediaType: "video",
-    video: mockVideoAsset(),
-  },
-  learnMoreBlock2: {
-    title: "learn more block 2",
-    bodyPortableText: portableTextFromString("block 2 text"),
-    alignMedia: "left",
-    mediaType: "image",
-    image: mockImageAsset("block2"),
-  },
-  seo: mockSeo(),
-};
+import { mockPosts } from "./index.test";
+import { testPlanALessonPageData } from "./lesson-planning.fixture";
 
-const testInternalLessonElementsCta: CTA = {
-  linkType: "internal",
-  label: "internal cta",
-  internal: { contentType: "aboutCorePage.whoWeAre", id: "1" },
-};
+import CMSClient from "@/node-lib/cms";
+import { BlogPostPreview } from "@/common-lib/cms-types";
 
+jest.mock("@/node-lib/cms");
+
+const mockCMSClient = CMSClient as jest.MockedObject<typeof CMSClient>;
+
+const testPlanningPageData = testPlanALessonPageData;
 const getPageData = jest.fn(() => testPlanningPageData);
 
 const render = renderWithProviders();
 
 describe("pages/lesson-planning.tsx", () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-    jest.resetModules();
-    jest.mock("../../../src/node-lib/cms/", () => ({
-      __esModule: true,
-      default: {
-        planningPage: jest.fn(getPageData),
-      },
-    }));
-  });
-
-  it("Renders correct title ", () => {
-    render(<PlanALesson pageData={testPlanningPageData} />);
-
-    expect(screen.getByRole("heading", { level: 1 }).textContent).toBe(
-      "Planning title",
+  it("Renders header hero component", () => {
+    const { getByRole } = render(
+      <PlanALesson pageData={testPlanningPageData} posts={mockPosts} />,
     );
+    expect(getByRole("heading", { name: "test" })).toBeInTheDocument();
   });
 
-  describe.skip("SEO", () => {
-    it("renders the correct SEO details", () => {
+  it("Renders a nav", () => {
+    render(<PlanALesson pageData={testPlanningPageData} posts={mockPosts} />);
+    const nav = screen.getByRole("navigation", {
+      name: "plan a lesson contents",
+    });
+    expect(screen.getAllByText("Contents")).toHaveLength(2);
+    expect(nav).toBeInTheDocument();
+  });
+
+  it("applies correct margin-bottom size based on section position if its a form block", () => {
+    render(<PlanALesson pageData={testPlanningPageData} posts={mockPosts} />);
+
+    const sections = screen.getAllByTestId("lesson-section");
+
+    expect(sections[0]).toHaveStyle("margin-bottom: 5rem");
+    expect(sections[1]).toHaveStyle("margin-bottom: 5rem");
+
+    expect(sections[sections.length - 1]).toHaveStyle("margin-bottom: 2rem");
+  });
+
+  it("applies correct margin-bottom size based on section position if its a content block", () => {
+    render(
+      <PlanALesson
+        pageData={{
+          ...testPlanningPageData,
+          content: testPlanALessonPageData.content.reverse(),
+        }}
+        posts={mockPosts}
+      />,
+    );
+
+    const sections = screen.getAllByTestId("lesson-section");
+
+    expect(sections[0]).toHaveStyle("margin-bottom: 5rem");
+    expect(sections[1]).toHaveStyle("margin-bottom: 5rem");
+
+    expect(sections[sections.length - 1]).toHaveStyle("margin-bottom: 2rem");
+  });
+
+  it("Renders the header hero with optional props", () => {
+    render(
+      <PlanALesson pageData={testPlanALessonPageData} posts={mockPosts} />,
+    );
+
+    expect(
+      screen.getByAltText(
+        `${testPlanALessonPageData.hero.author.name} profile picture`,
+      ),
+    ).toBeInTheDocument();
+    expect(
+      screen.getByAltText(testPlanALessonPageData.hero.image?.altText ?? ""),
+    ).toBeInTheDocument();
+  });
+  it("Renders the header hero without author", () => {
+    render(
+      <PlanALesson
+        pageData={{
+          ...testPlanALessonPageData,
+          hero: {
+            ...testPlanALessonPageData.hero,
+            author: { ...testPlanALessonPageData.hero.author, image: null },
+          },
+        }}
+        posts={mockPosts}
+      />,
+    );
+
+    expect(
+      screen.queryByAltText(
+        `${testPlanALessonPageData.hero.author.name} profile picture`,
+      ),
+    ).not.toBeInTheDocument();
+    expect(
+      screen.queryByAltText(testPlanALessonPageData.hero.image?.altText ?? ""),
+    ).not.toBeInTheDocument();
+  });
+
+  it("Renders the author title if it exists", () => {
+    render(
+      <PlanALesson
+        pageData={{
+          ...testPlanALessonPageData,
+          hero: {
+            ...testPlanALessonPageData.hero,
+            author: {
+              id: "13719a7b-4f00-4816-883a-8aded2a5c703",
+              name: "Katie Marl",
+              image: {
+                asset: {
+                  _id: "image-586258ca4b3b23d1a6fc47979841e5a5eb3dc36c-320x256-png",
+                  url: "https://cdn.sanity.io/images/cuvjke51/production/586258ca4b3b23d1a6fc47979841e5a5eb3dc36c-320x256.png",
+                },
+                hotspot: null,
+              },
+              bioPortableText: null,
+              socials: null,
+            },
+          },
+        }}
+        posts={mockPosts}
+      />,
+    );
+
+    expect(
+      screen.queryByText("Primary Curriculum Design Lead"),
+    ).not.toBeInTheDocument();
+  });
+
+  it("renders a hero image with alt text", () => {
+    render(
+      <PlanALesson pageData={testPlanALessonPageData} posts={mockPosts} />,
+    );
+
+    const heroImage = screen.getByAltText("alt text hero").closest("img");
+    expect(heroImage).toBeInTheDocument();
+  });
+
+  describe("SEO", () => {
+    it.skip("renders the correct SEO details", () => {
       const { seo } = renderWithSeo()(
-        <PlanALesson pageData={testPlanningPageData} />,
+        <PlanALesson pageData={testPlanningPageData} posts={mockPosts} />,
       );
 
       expect(seo).toEqual({});
@@ -117,83 +155,47 @@ describe("pages/lesson-planning.tsx", () => {
   });
 
   describe("getStaticProps", () => {
+    const mockPost = {
+      id: "1",
+      title: "Some blog post",
+      slug: "some-blog-post",
+      date: new Date("2022-12-01"),
+      category: { title: "Some category", slug: "some-category" },
+    } as BlogPostPreview;
+
+    beforeEach(() => {
+      jest.clearAllMocks();
+      jest.resetModules();
+
+      mockCMSClient.planALessonPage.mockResolvedValue(testPlanningPageData);
+      mockCMSClient.blogPosts.mockResolvedValue([]);
+      mockCMSClient.webinars.mockResolvedValue([]);
+    });
+
     it("Should not fetch draft content by default", async () => {
-      const { getStaticProps } = await import("../../pages/lesson-planning");
+      mockCMSClient.blogPosts.mockResolvedValueOnce([mockPost]);
       await getStaticProps({
         params: {},
       });
 
-      expect(getPageData).toHaveBeenCalledWith({
-        previewMode: false,
-      });
+      expect(mockCMSClient.blogPosts).toHaveBeenCalledWith(
+        expect.objectContaining({
+          previewMode: false,
+        }),
+      );
     });
 
     it("should return notFound when the page data is missing", async () => {
       getPageData.mockResolvedValueOnce(null as never);
+      mockCMSClient.planALessonPage.mockResolvedValueOnce(null);
 
       const { getStaticProps } = await import("../../pages/lesson-planning");
       const propsResult = await getStaticProps({
         params: {},
       });
-
       expect(propsResult).toMatchObject({
         notFound: true,
       });
-    });
-  });
-
-  describe("CTA links", () => {
-    it("should use pageData for the search CTA", () => {
-      render(<PlanALesson pageData={testPlanningPageData} />);
-      const searchCta = screen.getByRole("link", {
-        name: testPlanningPageData.stepsCTA.label,
-      });
-      expect(searchCta).toHaveAttribute("href", "/");
-    });
-    it("should use pageData for lesson elements CTA", () => {
-      render(<PlanALesson pageData={testPlanningPageData} />);
-      const lessonElementsCta = screen.getAllByRole("link", {
-        name: testPlanningPageData.lessonElementsCTA.label,
-      });
-      const externalLink =
-        testPlanningPageData.lessonElementsCTA.linkType === "external"
-          ? testPlanningPageData.lessonElementsCTA.external
-          : null;
-      if (!externalLink) {
-        throw new Error("This should never happen - check the test data");
-      }
-      expect(lessonElementsCta[0]).toHaveAttribute("href", externalLink);
-    });
-    it("should link to internal pages", () => {
-      render(
-        <PlanALesson
-          pageData={{
-            ...testPlanningPageData,
-            lessonElementsCTA: testInternalLessonElementsCta,
-          }}
-        />,
-      );
-      const lessonElementsCta = screen.getAllByRole("link", {
-        name: testInternalLessonElementsCta.label,
-      });
-      expect(lessonElementsCta[0]).toHaveAttribute(
-        "href",
-        "/about-us/who-we-are",
-      );
-    });
-    it("should open links in the same tab", () => {
-      render(
-        <PlanALesson
-          pageData={{
-            ...testPlanningPageData,
-            lessonElementsCTA: testInternalLessonElementsCta,
-          }}
-        />,
-      );
-      const lessonElementsCta = screen.getAllByRole("link", {
-        name: testInternalLessonElementsCta.label,
-      });
-      expect(lessonElementsCta[0]).not.toHaveAttribute("target", "_blank");
     });
   });
 });
